requirements.txt
python-telegram-bot==20.1
requests==2.27.1
apscheduler==3.8.1
flask==2.1.2
pytz==2022.1
